# openapi/protocol-v1.yaml
openapi: 3.1.0
info:
  title: unsea Protocol API v1
  version: 1.0.0-draft
  description: |
    新一代 WhatsApp 协议层接口（基于 Baileys 7.x）。
    本规范覆盖《Baileys 协议层接口封装与 2000 账号承载方案》§ 3 + 附录 A。

    ## 设计原则

    - 协议层只负责 ws / Noise / creds / 重连，业务编排在 Java 业务层
    - 所有状态查询都带 evidence（时间戳 + 来源）
    - 所有重连相关都对应明确的语义码（PROXY_FAILED / RATE_LIMITED / NEED_REAUTH ...）
    - 接口契约稳定，protobuf 内部码不外泄

    ---

    ## 业务场景导航

    ### 🔑 账号登录 / 授权（3 种方式 + 4 种导入）

    | 业务场景 | 接口 | 说明 |
    |---|---|---|
    | 首次绑定新号（pairing code） | `POST /v1/auth/pairing-code` | 用户在手机 WA "已登录设备" 里输入 8 位码；90s 不绑会自动释放 |
    | 首次绑定新号（扫码） | `POST /v1/auth/qrcode` | 用户用手机 WA 扫返回的二维码；180s 不绑会自动释放 |
    | 标准 Baileys JSON 导入（推荐） | `POST /v1/accounts/import/baileys-json` | 完整 creds + keys，导入即可上线 |
    | 全参数登录（30+ 字段） | `POST /v1/accounts/import/params` | 移植自 Apifox `paramsLogin` |
    | 六段参数登录 | `POST /v1/accounts/import/six` | 移植自 Apifox `sixLogin`，含 deviceIdentityKey |
    | 旧 JSON 登录 | `POST /v1/accounts/import/legacy-json` | 移植自 Apifox `jsonLogin`，base64 |
    | 批量导入（最多 500） | `POST /v1/accounts/import/batch` | 含冷启动节流，避免一次性灌爆 worker |
    | 短信注册新号 | （不支持） | WA 业务不开放此能力 |

    **关键概念**：换 IP 不需要重新授权（详见 § 4.3）。一次 pairing 后 creds 永久持有，重连用本地 creds 即可。

    ### 🚪 账号生命周期

    | 业务场景 | 接口 |
    |---|---|
    | 已绑账号上线（用已存 creds） | `POST /v1/accounts/{accountId}/online` |
    | 主动下线（保留 creds，可重新 online） | `POST /v1/accounts/{accountId}/offline` |
    | 退出并移除设备 | `POST /v1/accounts/{accountId}/logout` |

    ### 🏷️ 账号状态查询

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/messages/text:
    post:
      tags: [messages]
      summary: 发送文本消息
      description: "Baileys `sendMessage(jid, { text })`"
      operationId: messageText
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TextMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
        '503':
          description: 账号不可用（OFFLINE / NEED_REAUTH / restriction active）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/messages/image:
    post:
      tags: [messages]
      summary: 发送图片
      description: "Baileys `sendMessage(jid, { image, caption })`"
      operationId: messageImage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/audio:
    post:
      tags: [messages]
      summary: 发送音频（含 ptt 语音消息）
      description: "Baileys `sendMessage(jid, { audio, mimetype, ptt })`"
      operationId: messageAudio
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AudioMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/video:
    post:
      tags: [messages]
      summary: 发送视频
      description: "Baileys `sendMessage(jid, { video, caption })`"
      operationId: messageVideo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VideoMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/document:
    post:
      tags: [messages]
      summary: 发送文档
      description: "Baileys `sendMessage(jid, { document, mimetype, fileName })`"
      operationId: messageDocument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/location:
    post:
      tags: [messages]
      summary: 发送地理位置
      description: "Baileys `sendMessage(jid, { location })`"
      operationId: messageLocation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/contact-card:
    post:
      tags: [messages]
      summary: 发送名片（VCARD）
      description: "Baileys `sendMessage(jid, { contacts })`"
      operationId: messageContactCard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactCardMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/link:
    post:
      tags: [messages]
      summary: 发送链接（带预览）
      description: 等同 text 消息，协议层自动抓取 URL 预览
      operationId: messageLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LinkMessageBody' }
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/reaction:
    post:
      tags: [messages]
      summary: 给消息加 / 移除表情反应
      description: "Baileys `sendMessage(jid, { react: { text, key } })`"
      operationId: messageReaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReactionMessageBody' }
      responses:
        '200':
          description: 反应已发送
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/delete:
    post:
      tags: [messages]
      summary: 撤回消息
      description: "Baileys `sendMessage(jid, { delete: key })`"
      operationId: messageDelete
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteMessageBody' }
      responses:
        '200':
          description: 撤回成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
        '403':
          description: 超过撤回时限（WA 默认 ~2h）或消息不属于当前账号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/messages/forward:
    post:
      tags: [messages]
      summary: 转发消息
      description: "Baileys `sendMessage(jid, { forward: msg })`"
      operationId: messageForward
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForwardMessageBody' }
      responses:
        '200':
          description: 转发成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }
  /v1/messages/read:
    post:
      tags: [messages]
      summary: 标记消息已读
      description: "Baileys `readMessages([key])`"
      operationId: messageRead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReadMessagesBody' }
      responses:
        '200':
          description: 已标记
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
...
  /v1/messages/{messageId}/download:
    post:
      tags: [messages]
      summary: 按需下载消息媒体
      description: |
        Baileys `downloadMediaMessage(msg, 'buffer')`。
        协议层只负责按需下载 WhatsApp 媒体并返回 `base64` 或 `stream`。
        长期存储、对象存储上传、CDN URL 和生命周期由业务层负责。
      operationId: messageDownload
      parameters:
        - in: path
          name: messageId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DownloadMediaBody' }
      responses:
        '200':
          description: 下载成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DownloadMediaResult' }
        '404':
          description: 消息不存在或媒体已过期（WA 服务端 14 天）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/messages/status:
    post:
      tags: [messages]
      summary: 发送 status（动态）广播
      description: |
        Baileys `sendMessage('status@broadcast', content, { statusJidList })`。
        content 可以是文本/图片/视频。statusJidList 限定可见联系人。
      operationId: messageStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusBroadcastBody' }
      responses:
        '200':
          description: 动态已发布
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageSendResult' }

  # ──────────────────── A.7 联系人 / 聊天（占位） ────────────────────
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        evidence: { $ref: '#/components/schemas/Evidence' }
        details:
          type: object
          additionalProperties: true
          description: |
            错误详情。`NOT_OWNER` 时包含 accountId/currentWorkerId/ownerWorkerId/ownerEndpoint，
            业务层应刷新 owner 后直连 ownerEndpoint 重试一次。
    Evidence:
      type: object
      description: 状态来源证据（详见 § 4.7.1 三层证据模型）
      properties:
        wsOpen: { type: boolean }
        connectionField:
          type: string
          enum: [connecting, open, close]
        lastDateRecv: { type: string, format: date-time }
        lastPingAckAt: { type: string, format: date-time }
        ageMs: { type: integer }
        keepAliveIntervalMs: { type: integer, example: 30000 }
        slotReleased:
          type: boolean
          description: true 表示该账号当前不占 worker 在线容量
        reason:
          type: string
          nullable: true
          description: slotReleased 或状态变更原因
        updatedAt:
          type: string
          format: date-time
          description: runtime 状态更新时间
...
    MediaInput:
      type: object
      description: 媒体输入。三种方式三选一。
      properties:
        url:
          type: string
          format: uri
          description: HTTPS URL，协议层服务器自己下载
        base64:
          type: string
          description: base64 编码的二进制（不带 data:image 前缀）
        path:
          type: string
          description: 协议层本地路径（仅运维场景，业务侧请勿用）
        mimetype:
          type: string
          example: "image/jpeg"
    MessageKey:
      type: object
      required: [remoteJid, fromMe, id]
      properties:
        remoteJid:
          type: string
          example: "8617600627277@s.whatsapp.net"
        fromMe: { type: boolean }
        id:
          type: string
          example: "3EB0A5B6C7D8E9F0"
        participant:
          type: string
          nullable: true
          description: 群消息时为发送者 jid
    ContextInfo:
      type: object
      properties:
        quotedMessage:
          $ref: '#/components/schemas/MessageKey'
        mentionedJid:
          type: array
          items: { type: string }
        forwardingScore:
          type: integer
          description: 转发次数计数
    MessageSendResult:
      type: object
      required: [messageId, key]
      properties:
        messageId: { type: string }
        key: { $ref: '#/components/schemas/MessageKey' }
        timestamp: { type: integer, description: unix timestamp }
        status:
          type: string
          enum: [pending, server_ack, delivery_ack, read, played]
          description: 发送时通常返回 pending 或 server_ack；后续状态走 message.ack 事件
        recipients:
          type: array
          description: 群消息每个收件人的 ack（可选）
          items:
            type: object
            properties:
              jid: { type: string }
              status: { type: string }

    # ──── 消息发送 body ────
    TextMessageBody:
      type: object
      required: [accountId, jid, text]
      properties:
        accountId: { type: string }
        jid:
          type: string
          example: "8617600627277@s.whatsapp.net"
        text:
          type: string
          maxLength: 4096
        contextInfo:
          $ref: '#/components/schemas/ContextInfo'
        ephemeralExpirationSec:
          type: integer
          description: 阅后即焚 TTL；省略则跟群/会话设置
    ImageMessageBody:
      type: object
      required: [accountId, jid, image]
      properties:
        accountId: { type: string }
        jid: { type: string }
        image: { $ref: '#/components/schemas/MediaInput' }
        caption:
          type: string
          maxLength: 1024
        viewOnce: { type: boolean, default: false }
        contextInfo: { $ref: '#/components/schemas/ContextInfo' }
    AudioMessageBody:
      type: object
      required: [accountId, jid, audio]
      properties:
        accountId: { type: string }
        jid: { type: string }
        audio: { $ref: '#/components/schemas/MediaInput' }
        ptt:
          type: boolean
          default: false
          description: true=语音消息（push-to-talk），false=普通音频文件
        mimetype:
          type: string
          example: "audio/ogg; codecs=opus"
    VideoMessageBody:
      type: object
      required: [accountId, jid, video]
      properties:
        accountId: { type: string }
        jid: { type: string }
        video: { $ref: '#/components/schemas/MediaInput' }
        caption: { type: string, maxLength: 1024 }
        gifPlayback:
          type: boolean
          default: false
          description: true=作为 GIF 循环播放
        viewOnce: { type: boolean, default: false }
    DocumentMessageBody:
      type: object
      required: [accountId, jid, document, fileName]
      properties:
        accountId: { type: string }
        jid: { type: string }
        document: { $ref: '#/components/schemas/MediaInput' }
        fileName: { type: string, maxLength: 256 }
        mimetype: { type: string, example: "application/pdf" }
        caption: { type: string, maxLength: 1024 }
    LocationMessageBody:
      type: object
      required: [accountId, jid, degreesLatitude, degreesLongitude]
      properties:
        accountId: { type: string }
        jid: { type: string }
        degreesLatitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        degreesLongitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        name: { type: string }
        address: { type: string }
    ContactCardMessageBody:
      type: object
      required: [accountId, jid, contacts]
      properties:
        accountId: { type: string }
        jid: { type: string }
        contacts:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: object
            required: [displayName, vcard]
            properties:
              displayName: { type: string }
              vcard:
                type: string
                description: 标准 VCARD 3.0 字符串
    LinkMessageBody:
      type: object
      required: [accountId, jid, text]
      properties:
        accountId: { type: string }
        jid: { type: string }
        text:
          type: string
          description: 含 URL 的文本，协议层自动抓取预览
        generatePreview:
          type: boolean
          default: true
          description: false=只发文本不抓预览，省流量
    ReactionMessageBody:
      type: object
      required: [accountId, jid, targetKey, reaction]
      properties:
        accountId: { type: string }
        jid: { type: string }
        targetKey:
          $ref: '#/components/schemas/MessageKey'
        reaction:
          type: string
          description: 单个 emoji 字符；传空字符串 '' 表示移除 reaction
          example: "👍"
    DeleteMessageBody:
      type: object
      required: [accountId, jid, targetKey]
      properties:
        accountId: { type: string }
        jid: { type: string }
        targetKey:
          $ref: '#/components/schemas/MessageKey'
        forEveryone:
          type: boolean
          default: true
          description: true=撤回所有人；false=仅本地删除
    ForwardMessageBody:
      type: object
      required: [accountId, jid, sourceMessage]
      properties:
        accountId: { type: string }
        jid:
          type: string
          description: 目标会话
        sourceMessage:
          type: object
          description: 要转发的原始消息（可以是 message.received 事件 payload）
          additionalProperties: true
        forceForward:
          type: boolean
          default: false
          description: true=标记为"已转发"（Baileys 默认 false 不打转发标记）
    ReadMessagesBody:
      type: object
      required: [accountId, keys]
      properties:
        accountId: { type: string }
        keys:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/MessageKey' }
...
    DownloadMediaBody:
      type: object
      required: [accountId, message]
      properties:
        accountId: { type: string }
        message:
          type: object
          description: |
            完整 WA message 对象（含 key + message body）。
            业务侧从 `message.received` 事件 payload 直接取出来传过来。
            协议层用 Baileys downloadMediaMessage 解密 + fetch 媒体。
            **不要只传 key**：媒体 URL / mediaKey 等元数据在 message body 里，缺了无法解密。
          additionalProperties: true
        returnAs:
          type: string
          enum: [base64, url, stream]
          default: base64
          description: |
            base64: 直接返回二进制（小文件，<5MB）
            url: 兼容字段；协议层默认不上传对象存储，未配置存储时返回 501
            stream: HTTP chunked 流（仅 HTTP 协议层支持）
    DownloadMediaResult:
      type: object
      properties:
        mimetype: { type: string }
        sizeBytes: { type: integer }
        base64: { type: string, nullable: true }
        url:
          type: string
          format: uri
          nullable: true
          description: 兼容字段；协议层默认不上传对象存储，通常为 null
    StatusBroadcastBody:
      type: object
      required: [accountId, content, statusJidList]
      properties:
        accountId: { type: string }
        content:
          type: object
          description: 内容支持文本/图片/视频，结构同对应消息类型
          oneOf:
            - $ref: '#/components/schemas/TextMessageBody'
            - $ref: '#/components/schemas/ImageMessageBody'
            - $ref: '#/components/schemas/VideoMessageBody'
        statusJidList:
          type: array
          description: 可见的联系人 jid 列表（status@broadcast 限定可见范围）
          items: { type: string }

    # ──── Business catalog / product ────
...